# api/openapi.yaml
openapi: 3.0.3
info:
  title: Open ID Connect App
  description: |-
    This is the OIDC OpenID API documentation.
  termsOfService: https://localhost:1234
  contact:
    email: morning.night.dream@example.com
  license:
    name: MIT
  version: 0.0.1
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:1234
tags:
  - name: IdP
    description: Identity Provider
  - name: OP
    description: OpenID Provider
  - name: RP
    description: Relying Party
paths:
  /idp/signup:
    post:
      tags:
        - IdP
      summary: Sign Up
      description: Sign Up
      operationId: idpSignUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdPSignUpRequestSchema'
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
  /id/signin:
    post:
      tags:
        - IdP
      summary: Sign In
      description: Sign In
      operationId: idpSignIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdPSignInRequestSchema'
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
  /op/.well-known/openid-configuration:
    get:
      tags:
        - OP
      summary: OpenID Provider Configuration
      description: OpenID Provider Configuration
      operationId: opOpenIDConfiguration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPOpenIDConfigurationResponseSchema'
        '500':
          description: Internal Server Error
  /op/authorize:
    get:
      tags:
        - OP
      summary: Authentication Request
      description: Authentication Request
      operationId: opAuthorize
      parameters:
        - name: response_type
          in: query
          required: true
          description: response_type
          schema:
            type: string
            enum:
              - code
              - id_token
              - token
              - code id_token
              - code token
              - id_token token
              - code id_token token
        - name: scope
          in: query
          required: true
          description: scope
          schema:
            type: string
            enum:
              - openid
              - profile
              - email
              - address
              - phone
              - offline_access
        - name: client_id
          in: query
          required: true
          description: client_id
          schema:
            type: string
            format: uri
            example: client_id
        - name: redirect_uri
          in: query
          required: true
          description: http://localhost:1234/rp/auth/callback
          schema:
            type: string
            format: uri
            example: http://localhost:1234/rp/auth/callback
        - name: state
          in: query
          required: false
          description: state
          schema:
            type: string
            example: state
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:1234/rp/auth/callback
        '500':
          description: Internal Server Error
  /op/login/view:
    get:
      tags:
        - OP
      summary: OP Login
      description: OP Login
      operationId: opLoginView
      parameters:
        - name: auth_request_id
          in: query
          required: true
          description: auth request id
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
  /op/login:
    post:
      tags:
        - OP
      summary: OP Login
      description: OP Login
      operationId: opLogin
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
  /op/callback:
    get:
      tags:
        - OP
      summary: OP Callback
      description: OP Callback
      operationId: opCallback
      parameters:
        - name: id
          in: query
          required: true
          description: id
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:1234/rp/callback
        '500':
          description: Internal Server Error
  /op/token:
    post:
      tags:
        - OP
      summary: Token Request
      description: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#TokenRequest
      operationId: opToken
      parameters:
        - name: grant_type
          in: query
          description: grant_type
          schema:
            type: string
            enum:
              - authorization_code
              - refresh_token
              - client_credentials
              - password
              - urn:ietf:params:oauth:grant-type:device_code
        - name: code
          in: query
          description: code
          schema:
            type: string
        - name: redirect_uri
          in: query
          description: http://localhost:1234/rp/callback
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPTokenResponseSchema'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: error
                    example: invalid_request
        '500':
          description: Internal Server Error
  /op/userinfo:
    get:
      tags:
        - OP
      summary: UserInfo Request
      description: https://openid.net/specs/openid-connect-core-1_0.html#UserInfo
      operationId: opUserinfo
      security:
        - Bearer: []      
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPUserInfoResponseSchema'
        '500':
          description: Internal Server Error
  /rp/login:
    get:
      tags:
        - RP
      summary: RP Login
      description: RP Login
      operationId: rpLogin
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:1234/op/authorize
        '500':
          description: Internal Server Error
  /rp/callback:
    get:
      tags:
        - RP
      summary: RP Callback
      description: RP Callback
      operationId: rpCallback
      parameters:
        - name: code
          in: query
          required: true
          description: code
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        - name: state
          in: query
          required: true
          description: state
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    IdPSignUpRequestSchema:
      type: object
      properties:
        id:
          type: string
          description: id
          example: id
        password:
          type: string
          description: password
          example: password
      required:
        - id
        - password
    IdPSignInRequestSchema:
      type: object
      properties:
        id:
          type: string
          description: id
          example: id
        password:
          type: string
          description: password
          example: password
      required:
        - id
        - password
    OPOpenIDConfigurationResponseSchema:
      type: object
      properties:
        issuer:
          type: string
          description: http://localhost:1234/op
          format: uri
          example: http://localhost:1234/op
        authorization_endpoint:
          type: string
          description: http://localhost:1234/op/authorize
          format: uri
          example: http://localhost:1234/op/authorize
        token_endpoint:
          type: string
          description: http://localhost:1234/op/token
          format: uri
          example: http://localhost:1234/op/token
        userinfo_endpoint:
          type: string
          description: http://localhost:1234/op/userinfo
          format: uri
          example: http://localhost:1234/op/userinfo
        revocation_endpoint:
          type: string
          description: http://localhost:1234/op/revoke
          format: uri
          example: http://localhost:1234/op/revoke
      required:
        - issuer
        - authorization_endpoint
        - token_endpoint
        - userinfo_endpoint
        - revocation_endpoint
    OPLoginRequestSchema:
      type: object
      properties:
        id:
          type: string
          description: id
          example: id
        username:
          type: string
          description: username
          example: username
        password:
          type: string
          description: password
          example: password
      required:
        - id
        - username
        - password
    OPTokenResponseSchema:
      type: object
      description: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#TokenResponse
      properties:
        access_token:
          type: string
          description: access_token
          example: access_token
        token_type:
          type: string
          description: token_type
          example: token_type
        refresh_token:
          type: string
          description: refresh_token
          example: refresh_token
        expires_in:
          type: integer
          description: expires_in
          example: 3600
        id_token:
          type: string
          description: id_token
          example: id_token
    OPUserInfoResponseSchema:
      type: object
      description: https://openid.net/specs/openid-connect-core-1_0.html#UserInfoResponse
      properties:
        sub:
          type: string
          description: sub
          example: sub
        name:
          type: string
          description: name
          example: name
        given_name:
          type: string
          description: given_name
          example: given_name
        family_name:
          type: string
          description: family_name
          example: family_name
        preferred_username:
          type: string
          description: preferred_username
          example: preferred_username
      required:
        - sub
        - name
        - given_name
        - family_name
        - preferred_username
    IDTokenSchema:
      type: object
      description: |
        https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#IDToken
        https://openid.net/specs/openid-connect-core-1_0.html#StandardClaims
      properties:
        iss:
          type: string
          description: issuer
          example: iss
        sub:
          type: string
          description: subject
          example: sub
        aud:
          type: string
          description: audience
          example: aud
        exp:
          type: integer
          description: expires_at
          example: 1687360973
        iat:
          type: integer
          description: issued_at
          example: 1687360973
        auth_time:
          type: integer
          description: auth_time
          example: 1687360973
        nonce:
          type: string
          description: nonce
          example: nonce
      required:
        - iss
        - sub
        - aud
        - exp
        - iat
        - auth_time
        - nonce